Revert "Colorisation"
