+++ dvkbuntu-easy-menu-qt_fr_FR.ts
@@ -1,0 +1,3 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!DOCTYPE TS>
+<TS version="2.1" language="dvkbuntu-easy-menu-qt_fr_FR"></TS>
