+++ main.cpp
@@ -1,4 +1,5 @@
 #include "mainwindow.h"
+
 #include <QApplication>
 #include <QWebEngineView>
 #include <QWebEngineSettings>
@@ -10,8 +11,7 @@ int main(int argc, char *argv[])
 
     MainWindow w;
     w.setWindowState(Qt::WindowMaximized);
-    w.showFullScreen();
-    //w.show();
+    w.show();
 
     return a.exec();
 }
