+++ mainwindow.h
@@ -38,10 +38,8 @@ public:
     QFont fontN;
     QFont fontD;
     QFont fontM;
-    QSize QS1;
-    int fSize1;
-    QSize QS2;
-    int fSize2;
+    QSize QS;
+    int fSize;
     int HEIGHT;
     QProcess *m_process = new QProcess;
     QProcessEnvironment env = QProcessEnvironment::systemEnvironment();
@@ -58,11 +56,6 @@ private:
     QProcess *email = new QProcess;
     QWebEngineView music;// = new QWebEngineView();
     bool play = false;
-    QString program;
-    QString PIDtxt;
-    QProcess WidFromPid;
-    WId myWinID;
-    int myPid;
 
 private slots:
     virtual bool eventFilter(QObject* watched, QEvent* event);
@@ -72,6 +65,7 @@ private slots:
     void on_Internet_clicked();
     void on_Music_clicked();
     void on_Discord_clicked();
+
 };
 
 #endif // MAINWINDOW_H
