--- src/resource.qrc
@@ -1,37 +1,0 @@
-<RCC>
-    <qresource prefix="/">
-        <file>../Images/0-Categorie/calculator.svg</file>
-        <file>../Images/0-Categorie/clipboard.svg</file>
-        <file>../Images/0-Categorie/envelope.svg</file>
-        <file>../Images/0-Categorie/globe-africa.svg</file>
-        <file>../Images/1-NavigateurWeb/arrow-left.svg</file>
-        <file>../Images/1-NavigateurWeb/arrow-right.svg</file>
-        <file>../Images/1-NavigateurWeb/home.svg</file>
-        <file>../Images/1-NavigateurWeb/print.svg</file>
-        <file>../Images/1-NavigateurWeb/star.svg</file>
-        <file>../Images/1-NavigateurWeb/sync-alt.svg</file>
-        <file>../Images/2-Calculatrice/cogs.svg</file>
-        <file>../Images/2-Calculatrice/euro-sign.svg</file>
-        <file>../Images/3-PriseDeNotes/copy.svg</file>
-        <file>../Images/3-PriseDeNotes/file-alt.svg</file>
-        <file>../Images/3-PriseDeNotes/file-export.svg</file>
-        <file>../Images/3-PriseDeNotes/file-import.svg</file>
-        <file>../Images/3-PriseDeNotes/file.svg</file>
-        <file>../Images/3-PriseDeNotes/pen.svg</file>
-        <file>../Images/3-PriseDeNotes/plus-square.svg</file>
-        <file>../Images/3-PriseDeNotes/print.svg</file>
-        <file>../Images/4-Courriel/envelope-open-text.svg</file>
-        <file>../Images/4-Courriel/envelope-open.svg</file>
-        <file>../Images/4-Courriel/envelope.svg</file>
-        <file>../Images/4-Courriel/print.svg</file>
-        <file>../Images/EasyMenu_Icone.svg</file>
-        <file>../Images/0-Categorie/music.svg</file>
-        <file>../Images/0-Categorie/discord.svg</file>
-        <file>../Images/0-Categorie/calculator-color.svg</file>
-        <file>../Images/0-Categorie/clipboard-color.svg</file>
-        <file>../Images/0-Categorie/discord-color.svg</file>
-        <file>../Images/0-Categorie/envelope-color.svg</file>
-        <file>../Images/0-Categorie/globe-africa-color.svg</file>
-        <file>../Images/0-Categorie/music-color.svg</file>
-    </qresource>
-</RCC>
